Minor fixes across codebase

Former-commit-id: 9086b1389c6ccd7ba62633728eae5410fe6b505b

+++ deeplearning4j-nlp-parent/deeplearning4j-nlp/src/main/java/org/deeplearning4j/models/word2vec/wordstore/inmemory/InMemoryLookupCache.java
@@ -290,17 +290,17 @@ public class InMemoryLookupCache implements VocabCache<VocabWord>,Serializable {
     }
 
     @Override
-    public synchronized void incrementDocCount(String word, int howMuch) {
+    public synchronized void incrementDocCount(String word, long howMuch) {
         docFrequencies.incrementCount(word, howMuch);
     }
 
     @Override
-    public synchronized void setCountForDoc(String word, int count) {
+    public synchronized void setCountForDoc(String word, long count) {
         docFrequencies.setCount(word, count);
     }
 
     @Override
-    public synchronized int totalNumberOfDocs() {
+    public synchronized long totalNumberOfDocs() {
         return numDocs;
     }
 
@@ -310,7 +310,7 @@ public class InMemoryLookupCache implements VocabCache<VocabWord>,Serializable {
     }
 
     @Override
-    public synchronized void incrementTotalDocCount(int by) {
+    public synchronized void incrementTotalDocCount(long by) {
         numDocs += by;
     }
